requirements.txt
astroid==2.4.2
colorama==0.4.4
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.5
pylint==2.6.0
PyPDF2==1.26.0
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
xlrd==1.2.0
